+++ src/app/home/page.js
@@ -3,14 +3,14 @@
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
 
-import { useFetchData } from "../../hooks/useFetchData";
+import { useFilteredBookmarks } from "../../hooks/useFilteredBookmarks";
 import { useAuthContext } from "../../context/AuthContext";
 
 import Card from "../../components/Card";
 import NoContents from "../../components/NoContents";
 
 export default function Page() {
-  const { filteredBookmarks, dataLoading } = useFetchData();
+  const { filteredBookmarks, dataLoading } = useFilteredBookmarks();
   const { currentUser, loading } = useAuthContext();
   const router = useRouter();
 
@@ -26,7 +26,7 @@ export default function Page() {
         <NoContents />
       ) : (
         <div className="flex-grow grid grid-cols-3 gap-x-4 gap-y-4 my-8 px-12 max-w-5xl mx-auto border-4">
-        {filteredBookmarks.map((bookmark) => {
+        {filteredBookmarks?.map((bookmark) => {
           return (
             <Card key={bookmark.id} id={bookmark.id} url={bookmark.url} title={bookmark.title} bookmarkTags={bookmark.tags} />
           )
