+++ src/hooks/useFetchData.js
@@ -1,9 +1,8 @@
 import axios from "axios";
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import useSWR from 'swr';
 
 import { useAuthContext } from "../context/AuthContext";
-import { useSearchContext } from "../context/SearchContext";
 
 const fetcher = async (url, currentUser) => {
   return currentUser?.getIdToken()
@@ -22,38 +21,18 @@ const fetcher = async (url, currentUser) => {
 };
 
 export function useFetchData() {
-  const { currentUser, loading } = useAuthContext();
+  const { currentUser } = useAuthContext();
   const { data: bookmarks, error } = useSWR(
     currentUser ? [`/api/v1/bookmarks`, currentUser] : null, 
     ([url, currentUser]) => fetcher(url, currentUser)
   );
-  const [filteredBookmarks, setFilteredBookmarks] = useState([]);
-  const [dataLoading, setDataLoading] = useState(false);
-  const { selectedTags, setSelectedTags } = useSearchContext();
 
   useEffect(() => {
-    if (!loading && bookmarks) {
-      setFilteredBookmarks(bookmarks);
-      setSelectedTags([]);
-      setDataLoading(true);
-    }
-  }, [bookmarks]);
-
-  useEffect(() => {
-    if (!selectedTags.length) {
-      setFilteredBookmarks(bookmarks);
-    } else {
-      const result = bookmarks.filter((bookmark) =>
-        selectedTags.every((selectedTag) =>
-          bookmark.tags.some((tag) => tag.name === selectedTag)
-        )
-      );
-      setFilteredBookmarks(result);
-    }
-  }, [selectedTags]);
+    if (error) console.log(`error message: ${error}`);
+  }, [error]);
 
   return {
-    filteredBookmarks,
-    dataLoading
+    bookmarks,
+    error
   };
 }
